Removes MIT Licensing moving forward
I want this project to be transparent, in the sense that it is subject to public scrutiny. I also want anyone who feels compelled to be able to contribute freely (via a pull request or issue) with minimal barriers. This is possible without granting everyone access to release their own clones of the software.

Of course, anyone is welcome to clone & build on previous versions of this software -- the proof-of-concept is free to all, and it can serve as the foundation / inspiration for a number of applications. I will also be willing to consider custom enterprise licensing in the future to integrate with & improve the software operations of any company who finds this useful... just message me!